Split the WebAuth module into a Schema and Context
This commit splits the monolithic WebAuth module in to a Schema
and a Context.

See https://hexdocs.pm/phoenix/contexts.html for more info.
